requirements.txt
streamlit
opencv-python
faiss-cpu
Pillow
transformers
torch
aiofiles
requests
